fix(#1212): API find file feature parity
