+++ lua/nvim-tree/actions/init.lua
@@ -11,6 +11,9 @@ function M.setup(opts)
   require("nvim-tree.actions.fs.remove-file").setup(opts)
   require("nvim-tree.actions.fs.copy-paste").setup(opts)
   require("nvim-tree.actions.tree-modifiers.expand-all").setup(opts)
+  require("nvim-tree.actions.tree.find-file").setup(opts)
+  require("nvim-tree.actions.tree.open").setup(opts)
+  require("nvim-tree.actions.tree.toggle").setup(opts)
 end
 
 return M
